Move draw_top_separator to shapes module
